Reorder plugins to make globalData available during config
