+++ packages/11ty/_plugins/globalData/index.js
@@ -36,27 +36,45 @@ const checkForDuplicateIds = function (data, filename) {
 }
 
 /**
+ * @todo replace with ajv schema validation
+ */
+const validatePublication = (publication) => {
+  try {
+    const { url } = publication
+    publication.url = url.endsWith('/') ? url : url + '/'
+    publication.pathname = new URL(publication.url).pathname
+  } catch (errorMessage) {
+    logger.error(
+      `Publication.yaml url property must be a valid url. Current url value: "${url}"`
+    )
+    throw new Error(errorMessage)
+  }
+  return publication
+}
+
+/**
  * Eleventy plugin to programmatically load global data from files
  * so that it is available to shortcode components.
  * Nota bene: data is loaded from a sub directory of the `input` directory,
  * distinct from the `eleventyConfig.dir.data` directory.
  */
-module.exports = {
-  configFunction: function(eleventyConfig, options = {}) {
-    const dir = path.resolve(eleventyConfig.dir.input, '_data')
-    // console.debug(`[plugins:globalData] ${dir}`)
-    const files = fs.readdirSync(dir)
-      .filter((file) => path.extname(file) !== '.md')
-    const parse = parser(eleventyConfig)
+module.exports = function(eleventyConfig, { inputDir }) {
+  const dir = path.resolve(inputDir, '_data')
+  // console.debug(`[plugins:globalData] ${dir}`)
+  const files = fs.readdirSync(dir)
+    .filter((file) => path.extname(file) !== '.md')
+  const parse = parser(eleventyConfig)
 
-    for (const file of files) {
-      const { name: key } = path.parse(file)
-      const value = parse(path.join(dir, file))
-      if (key && value) {
-        checkForDuplicateIds(value, file)
-        eleventyConfig.addGlobalData(key, value)
-      }
+  for (const file of files) {
+    const { name: key } = path.parse(file)
+    let value = parse(path.join(dir, file))
+    if (key === 'publication') {
+      value = validatePublication(value)
     }
-  },
-  initArguments: {}
+    if (key && value) {
+      checkForDuplicateIds(value, file)
+      eleventyConfig.addGlobalData(key, value)
+    }
+  }
+  return eleventyConfig.globalData
 }
